+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.5
+
+- Added binary executable which helps download and verify the latest Monero release binary archive.
+
+## 0.0.4
+
+- Example fix
+- native_toolchain_rust fix: remove dangling reference in build/hook.dart since 0.0.3.
+
 ## 0.0.3
 
 - native_toolchain_rust fixes: using native_toolchain_rust_mirror until its parent merges lib fixes.